requirements.txt
mattermostdriver
requests
python-gitlab
python-dotenv

coverage
